+++ CHANGES.rst
@@ -88,6 +88,10 @@ Bug Fixes
 
 - Restrict polynomial order value of Polynomial1D model component to be >= 0. [#3741]
 
+- When all data is unloaded from 1D Spectrum viewer, and new data is loaded, the unit setting in the unit conversion plugin from the initial
+  dataset are maintained. [#3746]
+
+
 Cubeviz
 ^^^^^^^
 - Fixed issue with initial model components not using spectral y axis unit. [#3715]
@@ -109,6 +113,8 @@ Specviz
 
 Specviz2d
 ^^^^^^^^^
+
+- Fix various bugs in the Line Analysis plugin, including some issues with spectral axis unit conversions and deletion of subsets. [#3746]
 
 4.3 (2025-05-28)
 ================